requirements.txt
python-telegram-bot[socks]
emoji
requests
emoji
ephem